pyproject.toml
[project]
name = "django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.8.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "drf-spectacular>=0.28.0",
]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
]
